Fix bugs in show error message
How this working? Basically there is a <pre> tag and a <textarea> and I need show up the errors when the mouse hover a <mark> tag with a error, the big problem is that I can't assign events to <mark> tags directly because it will not get fired because they are behind the <textarea> element on the screen, so I needed work with the position of the cursor inside the <textarea> and calculate the position of the <mark> tag based on it.

I added some data-* attributes to the mark tags when they are created, according to: https://html.spec.whatwg.org/#embedding-custom-non-visible-data-with-the-data-*-attributes.

There is a lot of garbage code will gonna be removed when the funcionality get approved by @strager.

Most of the logic are coded inside the `index.ejs.html`, because I had import order issues when tried to put in `demo.mjs`, I gonna figure out how to put there.