Improve show error messages boxs
Change position of the error message box based on the cursor position.

Fix error message box width when the error is too large.

Substitute the every() function call to simple js for loop.

Fix problem when error message box is showing up and the user change the text and the box still appearing. The same was happening when the cursor get out of the textarea.

Fix bug of the creation of the error message box infinitely if the cursor is always over the statement with error.