Refactor code
First of all remove two classes to the same element and put a id.

Change the function to remove error box message. Now instead of loop multiples box, just get the only  box and remove it.

Remove function to check if box exists, because a function to remove the box is called before check if.

Add event listeners, "input", "click" and "mouseout" to textarea to remove the error box message.