Refactor --config-file
The logic which implements --config-file is in main.cpp, making it hard
to test. Factor it into a configuration-loader module, wrap it in a
class, and test it.

This commit does change behavior. Before, all config files were loaded
before opening any input files. After, a config file is loaded just
before its input file is opened and read. This affects when error
messages are reported.