Sublime Text: Fix bug in location of invalid regions
When some 4-byte character or, in other words, some character with
the size of two UTF-16 code units (for example, emojis) comes before
some error region or warning region, that region is misplaced against
the text of the JavaScript code file. The solution was to detect these
characters in the text and replace them with whitespace before sending
the text to quick-lint-js.

This commit also fixes an implementation bug where instead of passing
the number of bytes of text to quick-lint-js, it passes the number of
characters.