Sublime Text: Update previous bug fix solution
This commit changes the string that replaces the 4-byte characters.
This string is replaced from whitespace to <Not a Character>, because
if the replacement is done by the whitespace character quick-lint-js
will report, for example, "missing variable name" instead of "not
character allowed in identifiers". The idea is as I believe that all
4-byte characters are not valid to be used as identifiers of variables
or functions in JavaScript, so I should replace them with a character
that is also not valid, for quick-lint-js to interpret correctly the
text being sent. In this case, just to ensure your understanding, if
I replace it with whitespace quick-lint-js will interpret such character
as the absence of something, if I replace it with <Not a Character>
quick-lint-js will interpret such character as the presence of something
invalid for JavaScript language variables identifiers and function
identifiers.

This commit also adds some comments to the code to improve the
understanding of those reading the code.