More validation errors
Fixed validation errors in /license page where paragraph tag is not allowed to
be a children of an ordered list, and added lang attribute to html element (this
attribute was also added to /hiring and /benchmarks pages) #335. In /benchmarks page, we still get the following validation error `Error CSS: color: 70% is not a valid color 3 or 6 hexadecimals numbers`, which seems to be a validator bug. I have opened up an issue with validator, which can be tracked [here](https://github.com/validator/validator/issues/1160)