Refactor: pass doc version as JSON string
I want to save a document's version string and use it later. We can't
save values from simdjson because the lifetime of the version string
will outlast the lifetime of the simdjson parser.

Refactor lsp_javascript_linter to work with stored version JSON strings
(instead of a simdjson value).

This commit should not change behavior.