Add function to append byte_buffer-s
I want to refactor lsp_pipe_writer to support non-blocking writes on
Windows. This requires sending data to another thread. To make this
easy, I want to concatenate byte_buffer-s together.

Create function to move the chunks of a byte_buffer to the end of a
byte_buffer_iovec.

The new function is unused, so this commit should not change behavior.