Refactor pipe_reader; rename to event_loop
pipe_reader is blocking, but we need to implement multiplexing
(poll/WaitForMultipleObjects).

Refactor pipe_reader to take instructions from its derived class rather
than holding ownership of the readable pipe and the parser. This will
make it easier to extend pipe_reader (now event_loop) to work with a
writable pipe and with an inotify/kqueue/ReadDirectoryChanges handle.

This commit should not change behavior.