Use poll() on POSIX in event_loop
On POSIX platforms (Linux and macOS), the event_loop class blocks in
read(). This forces us to call write() on a separate thread. Make the
read pipe non-blocking and use poll() to wait for data to arrive. This
will let us multiplex read() and write() in the future.

On Windows, keep event_loop as-is (blocking ReadFile).

This commit should not change behavior.