Added Swedish localisation for quick-lint-js
Submitting my pull request Swedish translation/localisation for quick-lint-js. Felt as a something I would be able to start with as a  contribution to the project, even though the value might not be of the greatest significance.

Assuming most people and/or maintainers don't speak Swedish I will say that I tried to stay as close as I could to the message and context given by the English messages.