Rewrite event_loop using I/O completion ports (Windows)
For the LSP server, on Windows, we need an event pump based on
GetQueuedCompletionStatus (I/O completion ports) in order to support
filesystem change notifications (which will be used for detecting
changes to config files). Rewrite event_loop to use I/O completion
ports, moving the blocking pipe read call to a separate thread.

This commit should not change behavior.