Add opaque token for config change notifications
The LSP server needs to associate a changed config to its lsp_document
in order to re-lint open .js files. One solution is to maintain a map
from configuration_loader's input_path to a file URI (key to
lsp_server::documents_). This is kinda gross.

As a (hopefully) simpler solution, add a token pointer
configuration_change-s.

The token is currently only used in tests. Therefore, this commit should
not change behavior.