[10/n] Configurable globals: VS Code LSP: lint .js files on config changes
In our LSP-based VS Code extension, tell the quick-lint-js LSP server
about changes to quick-lint-js.config files. Do this by telling VS Code
that quick-lint-js.config files are JSON files, and that changes to JSON
files should be communicated to the quick-lint-js LSP server.

The net effect of this commit is that changes to quick-lint-js.config
re-lint affected .js files.