Create error message pop out box in website demo
In the website of this repository exists a `demo` of the linter where you can try some code and the linter will mark the sentence when you make some mistake. 

This PR have intent to create a feature that kind of pop out up a message describing what the error is about with respective `code` correspondent to the error, when your mouse hover the sentence with errors.

The implementation of this feature created some issues in the tests caused by the creation of the `data-` attributes inside the `<mark>` tags. Some of the issues were:

- The tests were not assigning `data-` attributes when instantiating the `<mark>` tags, so this tags were coming with undefined value assigned to them, so the testes were getting failed because the expect values in the tests have changed.
- The `code_block` function, which besides other things, checks if exists `BOM` in some HTML code passed to it, broke because the structure of the `<mark>` tag has changed (now having `data-` attributes inside) and this function was making string comparison, so was necessary change this string checking way to a `DOM` manner. With `DOM` manipulation is possible to check the content of the hole instead hard code string checking.

So beyond the implementation of the box error message feature this PR includes the fixes for this issues.