Website: smoothly select rows in bar chart
If you scrub your cursor slowly down the benchmark chart, the benchmark
table appears to flicker. This is because, if the cursor is between two
chart rows, no row is selected, so every row of the table is saturated.

Fix this issue by removing the cursor dead zones between chart rows.

This commit in theory shouldn't change the visuals, but it does move
things around slightly (at least in Firefox). As best as I could, I
tried to change the hit boxes while keeping everything the same size as
before.