Sublime Text: Make loading the shared library less troublesome.
Before this commit, if there is a plugin update, there would be a folder
with the shared library remaining from the previous plugin version.

Before a plugin update:
|-- quick-lint-js-v020-linux_x86_64.sublime-package
|-- quick-lint-js-v020-linux_x86_64.sublime-package
    |-- libquick-lint-js-lib.so

After a plugin update:
|-- quick-lint-js-v100-linux_x86_64.sublime-package
|-- quick-lint-js-v020-linux_x86_64.sublime-package
    |-- libquick-lint-js-lib.so
|-- quick-lint-js-v100-linux_x86_64.sublime-package
    |-- libquick-lint-js-lib.so

After this commit, if there is a plugin update, we won't have a folder
with the shared library remaining from the prior plugin version.

Before a plugin update:
|-- quick-lint-js-v020-linux_x86_64.sublime-package
|-- quick-lint-js
    |-- libquick-lint-js-lib.so

After a plugin update:
|-- quick-lint-js-v100-linux_x86_64.sublime-package
|-- quick-lint-js
    |-- libquick-lint-js-lib.so