Work around linker crash
When generating a link map, BFD LD version 2.26.1 crashes demangling the
following symbol:

    _ZSt7forwardIOZN13quick_lint_js11result_baseINS0_13padded_stringEJNS0_18read_file_io_errorENS0_19posix_file_io_errorEEEC4IJS3_EEEONS0_6resultIS2_JDpT_EEEEUlOT_E_ESD_RNSt16remove_referenceISC_E4typeE

(The above symbol doesn't exist in quick-lint-js now. It would have
existed in a future commit without this commit.)

It looks like the linker doesn't crash during demangling if we avoid
using a lambda. Rewrite result's conversion constructor to use a class
instead of a lambda, working around the crash.

This commit should not change behavior.