Don't automatically return default config in config_loader
I want configuration_loader's functions to return more than just a
pointer to the configuration object. To make this easier, make
configuration_loader's functions return nullptr instead of a default
config if a config file can't be found (or if an I/O error occurs).

This commit should not change behavior.