LSP: don't lint non-JS files as config files
Our LSP server treats anything which isn't JavaScript as if it was a
quick-lint-js.config file. This used to be fine, because we did nothing
interesting with such files. However, now we report diagnostics for
these files, which is misleading for, for example, .txt files and
non-quick-lint-js.config JSON files.

Only treat files named quick-lint-js.config or .quick-lint-js.config as
config files. Ignore all other documents.