CLI: add --path-for-config-search option
The Vim and Emacs editor plugins give quick-lint-js the JS code using
standard input (with --stdin) or a temp file. This means that
quick-lint-js looks for quick-lint-js.config in the wrong place: either
nowhere (for --stdin) or /tmp (for a temp file).

Make it possible for editor plugins to lint with the correct
quick-lint-js.config by adding the --path-for-config-search option.