Don't look for config in cwd for --stdin
If the user lints standard input (e.g. with --stdin), I think it's
surprising that quick-lint-js looks for a configuration file. --stdin is
often used by scripts and editors, but the current working directory is
a hidden input which silently changes quick-lint-js' behavior.

Avoid the lookup, and assume a default config for --stdin (unless
--config-file is also given).