CLI: fix --stdin reading config when it shouldn't
Commit 04cf5b36c5d9e0b37026797454f52d4231e2e7b1 changed how
configuration files are searched for --stdin. However, that commit
doesn't work correctly. In the CLI, file_to_lint::path is "<stdin>", not
nullptr. The tests don't do the same thing that the CLI does. This means
that config files are still searched in cwd. Oops!

Fix configuration_loader to check is_stdin explicitly. Include an
end-to-end test to be extra certain that --stdin doesn't use the
configuration file.