Vim: fix linting of new buffer
In Vim, if the user creates a new buffer with no name, quick-lint-js
searches for a configuration file starting in the parent of the current
working directory. This behavior is unintentional.

Change the Vim plugin to only pass the --path-for-config-search option
for buffers which represent files. THis means that new buffers will
always use the default configuration.