plugin/emacs: Don't use region intersection in Flymake while reporting
This is causing diagnostics to not be cleared sometimes (see #398), it
was first introduced for developement, because restarting Flymake
without changes in buffers would not delete old diagnostics, I think
it's fine if we not specify this option.

However I suspect this is a bug in Flymake.

Signed-off-by: wagner riffel <w@104d.net>
(cherry picked from commit 6a223f8d8e1693a287e40557600717f30bfe73ba)