wasm: recover when process crashes for another document
If multiple DocumentLinter-s are open with a single Process, and the
Process crashes for one DocumentLinter, the other DocumentLinter
continues to use the Process. This is a problem because the Process
might have a corrupted heap or some other problem. If a Process crashes,
we want to stop using it for any linting.

Before interacting with a Process, have DocumentLinter check if that
process has ever crashed for any DocumentLinter. If so, lint with a new
Process.

This commit should not affect the VS Code extension which creates a new
DocumentProcessManager per DocumentLinter.