wasm: create workspace to hold all open documents (C++ half)
In the VS Code extension, in order to have a .js file lint using an
open config file, we need to be able to correlate documents together. Do
this by creating a workspace abstraction to hold all the documents in
the C++ half.

A future commit will expose workspaces to the JavaScript half.

This commit should not change behavior.