VS Code: rewrite extension
The existing VS Code extension (plugin/vscode-wasm/) uses WebAssembly. I
need it to use native filesystem APIs, and this turned out to be a
headache.

Write a new VS Code extension which uses a Node.js addon instead of
WebAssembly for quick-lint-js' core.

The new extension is relatively inefficient. Improvements will be made
in future commits.

Many of the files, including the test suite, are copy-pasted from
plugin/vscode-wasm/.