VS Code: fix loading extension on Windows
The quick-lint-js-node-vscode Node.js addon tries to load a DLL called
node.exe. No such DLL exists in VS Code, so quick-lint-js-node-vscode
fails to load.

Redirect the loading of node.exe to the already-loaded code.exe. This is
how it's supposed to work, according to Electron's documentation:
https://github.com/electron/electron/blob/4f739d78371b5c91db010f7f931f74fdba34e11b/docs/tutorial/using-native-node-modules.md#a-note-about-win_delay_load_hook
(VS Code is based on Electron.)

This commit should only affect Windows and not other platforms.