Handle unexpected keywords in expressions
#407 

The new behavior is to end the expression for these keywords. It will not report an unexpected token error in an expression. 

Because the code continues parsing instead of crashing, there are known new issues with handling `catch`, `export`, `extends`, `finally` and `function` expressions. These issues occur when keywords are mentioned in unopened comment blocks.

For example:

```
*                  
* this function extends a string 
* /
```