VS Code: read config from disk after closing editor
When the user opens a quick-lint-js.config file, the editor's content is
used for linting .js files instead of the filesystem's. When the user
closes the quick-lint-js.config file, quick-lint-js doesn't reload the
config from disk when linting .js files.

When the user closes a config file, ake quick-lint-js forget the
previously-opened config file, and behave as if the file was never
opened for editing.