Fix parsing of for-in loop with var initializer
JavaScript allows var declarations in for-in loop to contain an
initializer. The initializer might contain unary ++, an arrow function,
or other things. quick-lint-js forgets that the 'in' operator is special
after parsing certain expressions, causing false positive errors in the
following code (for example):

    for (var x = ++y in []) {}
    for (var x = async (z) => y in []) {}

Propagate precedence::in_operator, fixing the false positives.