Fix missing config change notifications
If there are two watches with different tokens for the same watched
file, configuration_loader::refresh only reports one change with one of
the tokens. I don't know the consequences of this, but it seems wrong.

Fix configuration_loader::refresh to return changes for multiple tokens
even if the same path is watched.