macOS: don't crash on FS watch failures
On macOS, open() can fail if the file descriptor limit is exceeded.
quick-lint-js calls open() to watch directories for changes.

* Don't fail to open a config file if opening a directory fails.
* Show a warning message to the user (in their editor) if open() fails.
  Only show the first message to avoid spam.

Warning: This commit is probably pointless. If we recover from the
directory-watching open(), a subsequent file-reading open() is likely to
fail as well. We should probably clean up open watches on failure to
make room for more file descriptors.

This commit affects the LSP server and the VS Code extension.