Performance tracing implementation for vscode plugin
Issue Ref: #196
Implementation details:
Satisfies requirements to trace performance of linter and have the format specified in the original issue. Every new file using qljs is indexed via a file-ref number.

Added a new column called "event-id". This is meant to be a utility column that can store events that are occurring within qljs other than performance of parser. Currently, there are 2 different event-ids, a "n" which denotes a new file is opened using qljs while a "p" event denotes that a tracing event has occurred for a keystroke.