Fix leaks in lexer transactions
If a lexer transaction isn't committed or rolled back, memory is leaked.
Leaks can thus happen if QLJS_PARSER_UNIMPLEMENTED is called, since
the commit or rollback is skipped.

Instead of heap-allocating the buffering_error_reporter, embed it in the
lexer_transaction object. This means lexer_transaction needs to be
unmovable, but we don't need to move it anyway.