Add basic plumbing for parser options
In the future, we will need the parser to support conflicting syntaxes
(E.g. TypeScript vs Flow, or ES Modules vs scripts). The user will need
to configure quick-lint-js to get the best diagnostics.

Add a way to internally tweak the parser's treatment of 'await' at the
module scope.

In the future, this will be configured using quick-lint-js.config. For
now, this is only used in tests.