Rewrite logging; allow multiple loggers
I realized that the single-logger scheme isn't going to work. Store a
list of loggers instead of a single logger.

Also, centralize formatting so it doesn't need to be repeated for each
logger.

Also, because it's easiest to use a mutex anyway for the multi-logger
scheme, document that logger classes need not be thread-safe themselves.