Make 'delete x' not warn if x is global
'delete x' is a no-op if the module defines 'x' using 'var', 'let', etc.
However, if 'x' is a global variable, then 'delete x' has side effects.
quick-lint-js reports a warning in both cases, meaning it reports a
false positive in the latter case.

Make quick-lint-js report a warning on 'delete x' only if x is not a
global variable.