Detect and stop parsing on JSX
When quick-lint-js is run on code containing JSX, the parser and linter
barf horribly all over the code.

It'll take us a while to implement a full JSX parser. When JSX syntax is
detected, emit a friendly diagnostic and stop parsing to avoid
diagnostic spew.

For valid JavaScript code, this commit should not change behavior.