benchmark-lsp: rewrite in C++
benchmark-lsp is written in Haskell. This turned out to be a suboptimal
decision:

* lsp-test didn't do what we needed, so we had to effectively rewrite
  it.
* Criterion doesn't give us control over the number of iterations. This
  makes comparing different linters impossible because *N* is different
  for different linters. We would need to effectively rewrite Criterion
  to fix this problem.
* Haskell is foreign and scary for most people. (C++ is too, but less
  so.)

Rewrite benchmark-lsp in C++ to address these problems.

Notable differences:

* CLI options are very different.
* The output JSON format is very different.
* The "broken" config option is no longer supported.
* There is no HTML output format with pretty graphs.
* Each benchmark now has a warmup phase.

Because the CLI changed and the JSON format changed, the website's
instructions and scripts for the benchmarks are out of date. The website
will be fixed in a future commit.