Website: fade side nav hover background
On a wide monitor with a full-screened browser, the side bar nav's hover
effect extends to the left side of the screen. This is overkill. Add a
fade effect.

This commit changes the visuals, but does not change the click target.
(You can still click all the way to the left to activate the link.)