Website: fix position of mascot on hiring page
On pages without a tag line (such as /hiring/), Dusty (the mascot) is
too high on the page. This is because Dusty is positioned relative to
the tag line's grid cell. (A missing grid cell has a different size than
a grid cell with the tag line.)

Fix position of Dusty to be consistent regardless of the presence of the
tag line.