Website: tweak header some more
On mobile, the header is cramped because the 'quick-lint-js' title can
take two or three lines. Reduce the chance of word-wrapping the title by
moving Dusty below the title (but still left of the tag line) on smaller
screens.

To reduce word-wrapping of the tag line, shorten the copy.

Also, reduce the spacing between the title and the tag line.

Unrelatedly, remove the title's underline.