Use new code signing certificate
The existing code signing certificate is possibly compromised. I was
sloppy with it while trying to integrate code signing into GitHub CI.

Delete the existing code signing cert, replacing it with another.

This commit only affects signing of macOS and Windows executables.

The GPG key (used for signing Linux releases) was not exposed to GitHub
CI and remains unchanged.