benchmark-lsp: fix memory leak
When I used 'new' to allocate benchmarks, I thought it'd be fine.
However, I didn't think that we'd leak lots of memory if the user asks
for a high sample count.

Use unique_ptr to manage the memory of each benchmark so we can reuse
memory between samples.

This commit should not change behavior.