Rewrite compile-translations script in Go
For efficiency reasons, I want to encode our translations in a format
different than the GNU MO format. Rewrite our compile-translations from
Bash to Go to make it easier to encode our translations in a custom
format.

This commit should not change behavior. The script prints output
byte-identical to the original script (aside from a tweak to a comment).