add escape errors option
With '--escape-errors' option we can control the behaviour of escaped
errors. The default behaviour is 'auto' which uses 'isatty' to check
if the 'stderr' is a terminal or not and prints the escaped version
of errors only if it's a terminal.

Help page was also edited.

With this change tests will get unescaped errors by default. Also 2 new
tests where implemented for checking the behaviour of '--escape-errors'.