Rewrite generate-lex-unicode in Node.js
Python's unicodedata is install-specific, so it's hard to change which
Unicode version we use. Use a Unicode data library which allows
specifying the Unicode version [1]. In doing so, rewrite our script in
Node.js.

This commit should not change behavior. The generated file is
byte-identical (aside from changing the file path).

[1] https://www.npmjs.com/package/@unicode/unicode-13.0.0