CLI, VSCode and LSP error codes hyperlink to online docs
A new option was added (--diagnostic-hyperlinks) to control the
behaviour of CLI hyperlinks and it can accept (auto, always, never)
as arguments. Details are available in docs. Also Windows default
terminals don't support escape sequences and that's why 'auto'
behaves like 'never' in Windows.

This commit also improves some ZSH and Fish shells completions with
addition of the new mentioned option for Bash.