Vim: install docs and coc.nvim plugin
Our Vim plugin has some juicy features which are missing from the Debian
package. They're missing because we forgot to install them! Tell CMake
to install the Vim plugin documentation and the coc.nvim script.

Also, installers using vim-addons don't install all of the Vim plugin's
files either because the .yaml file for vim-addons is missing files.
List all the files for the Vim plugin in quick-lint-js.yaml.

This fixes `:help quick-lint-js` in DEbian, and probably fixes a
bunch of other stuff too.