Replace release docs with release script
Instead of manually following a text-based release process, automate the
release process with a script.

In reality, the script just prints the existing documentation then
prompts the user to continue to the next step. In the future, we can
extend the script to fully automate some steps.

Inspired by Dan Slimmon's do-nothing scripting idea:
https://blog.danslimmon.com/2019/07/15/do-nothing-scripting-the-key-to-gradual-automation/