Windows: optimize binary size; don't embed translations hash table
The MSVC toolchain is embedding translation_table_const_hash_table into
the .exe file. (Other toolchains seem to not embed the table into the
final executable.) The whole point of translation_table_const_hash_table
is that it should not be embedded.

Tweak translation_table_const_hash_table to be declared in a consteval
function. This forces the compiler to evaluate it at compile time only.
This seems to remove the table from the Windows .exe, shrinking the .exe
significantly.

This commit should not change behavior.