benchmark-lsp: fix waiting for Flow's diagnostics
The incremental-change-wait/express-router.js benchmark changes the
document in a way which shouldn't change diagnostics. When Flow receives
an incremental change, it first gives the client the old set of
diagnostics but with locations adjusted. Then, it re-parses, then
computes and sends truly updated diagnostics. The first set of
diagnostics confuses our benchmarks.

Fix our benchmarks to wait for the second set of diagnostics from Flow,
but only for benchmarks which send incremental changes. This makes the
benchmark measure what was intended.

This commit should not change the behavior of other benchmarks or change
the behavior of the incremental-change-wait/express-router.js benchmark
with other LSP servers.