Improve error recovery for 'await() => {await p;}'
When quick-lint-js sees something like 'await() => {}', it reports an
error (E0178) but parses the arrow function's body as if 'await' was
omitted. This leads to excessive errors if 'await' was supposed to be
'async' like the error suggests.

When reporting E0178, re-parse the arrow function as if it was an async
arrow function.