Improve error for 'await x => {await p;}' in non-async functions
In a non-async function, quick-lint-js parses the following code in a
suboptimal way:

    function f() {
      return await x => {await p;}
    }

    hello.js:2:10: error: 'await' is only allowed in async functions [E0162]
    hello.js:2:22: error: 'await' is only allowed in async functions [E0162]
    hello.js:2:28: warning: use of undeclared variable: p [E0057]

This is because 'await' is treated as an operator but doesn't trigger
E0178, causing the arrow function to be treated as non-async.

Treat the first 'await' as if it was 'async', similar to how the code
parses when 'f' is an async function. This produces better diagnostics:

    hello.js:2:10: error: 'await' cannot be followed by an arrow function; use 'async' instead [E0178]
    hello.js:2:28: warning: use of undeclared variable: p [E0057]