Improve error for 'await() => {await p;}' in non-async functions
In a non-async function, quick-lint-js parses the following code in a
confusing way:

    function f() {
      return await() => {await p;}
    }

    hello.js:2:22: error: 'await' is only allowed in async functions [E0162]
    hello.js:2:10: error: missing operator between expression and arrow function [E0063]
    hello.js:2:10: warning: use of undeclared variable: await [E0057]
    hello.js:2:28: warning: use of undeclared variable: p [E0057]

This is because the first 'await' is treated as a variable name and not
a keyword.

Treat the first 'await' as if it was 'async', similar to how the code
parses when 'f' is an async function. This produces better diagnostics:

    hello.js:2:10: error: 'await' cannot be followed by an arrow function; use 'async' instead [E0178]
    hello.js:2:28: warning: use of undeclared variable: p [E0057]