Update common-head.ejs.html
You are right ! my implementation was wrong.
I hosted the website in github pages and tested it with this online tool "https://seositecheckup.com/tools/canonical-tag-test".
I tried using the tool you recommended but it didn't work, it showed this error " Sitechecker bot is not allowed to crawl this website. "
but this i think this would work.