Avoid byte_buffer::append_copy with C string
byte_buffer::append_copy takes a string8_view, which is implicitly
convertible from a u8 string literal. This is slightly bad, because the
length of the string needs to be computed at run-time with strlen.

Force append_copy with a string literal to compute the size at compile
time.

This commit should not change behavior.