Refactor: visit class bodies promptly
When parse_expression encounters a class expression, it parses the body
of the class. During this parsing, visits are buffered via
buffering_visitor, and the buffering_visitor is stored in the class'
expression AST. When the AST is visited, the buffering_visitor's visits
are moved to the linter.

All this buffering is unnecessary. Instead of buffering, visit
immediately to the linter while parsing the class' body. This changes
the ordering of visits, but the ordering changes don't matter to
the linter.

Cleanup will happen in future commits.