Refactor: visit function bodies promptly
When parse_expression encounters a function, it parses the body of the
function. During this parsing, visits are buffered via
buffering_visitor, and the buffering_visitor is stored in the function's
expression AST. When the AST is visited, the buffering_visitor's visits
are moved to the linter.

All this buffering is unnecessary. Instead of buffering, visit
immediately to the linter while parsing the function's body. This
changes the ordering of visits, but the ordering changes don't matter to
the linter.

Cleanup will happen in future commits.