Docs: automatically install vim-addon-manager
When installing the Vim plugin from a Debian package, the user needs the
vim-addon-manager package to be installed. This is a dependency of the
Debian package, but installing with dpkg -i will not automatically
install the dependency.

Instead of suggesting dpkg -i, suggest apt-get install. This will prompt
the user to install dependencies as well.