Check for global_variable
For GCC 8, when compiling the release build, throws an error for
dereferencing null pointer. In debug build, we check for the nullity of
this pointer before dereferencing.

```c++
if (is_global_variable) {
  QLJS_ASSERT(!declaration);
} else {
  QLJS_ASSERT(declaration);
}
```

In release build, I think we skip these checks and the compiler assumes
that declaration is null pointer.