Improve diagnostic for if(x);else(y==z){}
quick-lint-js reports a bunch of confusing errors for the following
code:

    function lastName(name) {
      let parts = name.split();
      if (parts.length === 2) {
        return parts[1];
      } else (parts.length === 1) {  // line 9
        return name;
      } else {                       // line 11
        throw new Error(`unexpected name: ${name}`);
      }
    }
    // :9:35: error: missing arrow operator for arrow function [E0176]
    // :9:15: error: invalid function parameter [E0151]
    // :9:32: error: unexpected literal in parameter list; expected parameter name [E0159]
    // :11:9: error: 'else' has no corresponding 'if' [E0065]

This is caused by the `{` on line 9 being interpreted as the beginning
of an arrow function's body. This prevents E0184 (missing 'if' after
'else') from being reported, because the token after the `else`'s body
is `else` from line 11.

Disable the error recovery for E0176 (missing arrow operator for arrow
function) when parsing an else body that might look like an if
condition.

Error recovery is still imperfect for this example. E0176 is reported,
but so is E0065 ('else' has no corresponding 'if').