Improve error recovery for if(x){}else(y){}else{}
After parsing 'if(x){}else(y)', quick-lint-js ends the statement. This
means that 'if(x){}else(y){}else{}' is interpreted to have a dangling
else.

When error-recovering from 'else(y)', parse '{}' as its body, then try
parsing 'else' again. This makes the parser behave as if 'if' was
written after the first 'else'.