Refactor: split parse_template into untagged and tagged variants
parse_template feels weird to me. There are a few if statements with the
same boolean condition.

Refactor parse_template by splitting it into two functions:
parse_tagged_template and parse_untagged_template. This duplicates some
of the logic, but removes the ugly ifs.

This commit should not change behavior.