benchmark-lsp: embed benchmark-config.cpp
The benchmark-lsp tool requires the user to give the path to
benchmark-config.json. This is inconvenient and makes it hard to add new
features to the benchmark suite.

Embed benchmark-config.json's data into the benchmark-lsp tool (C++
code).

Because the config refers to files, the user still needs to specify the
directory which used to contain benchmark-config.json. Do this by
assuming the current working directory is the desired directory.