benchmark-lsp: fix miscounting of Deno diagnostics
The incremental-change-wait benchmarks change the document then wait for
diagnostics. Deno has three independent diagnostics sources. When each
source finishes its processing, Deno sends a diagnostics message to the
benchmark.

When Deno version 1.16.4 sends diagnostics messages, it reuses
diagnostics from the previous version of the document for sources which
have not finished processing. Our incremental-change-wait benchmark
works for this version of Deno because it waits for the next
diagnostics message after changing the document. That next diagnostics
message contains all the diagnostics from the previous version, so our
benchmark thinks Deno is done.

When Deno version 1.18.2 sends diagnostics messages, it does not use
diagnostics from any previous version of the document for sources which
have not finished processing.  This confuses our incremental-change-wait
benchmark because Deno will often send an empty list of diagnostics in
response to the document change.

TL;DR Deno version 1.18.2 is borken with our current scheme. Change the
benchmarks to wait for all diagnostics messages from Deno.